requirements.txt
asgiref==3.8.1
Django==4.2.20
django-ckeditor==6.7.2
django-js-asset==2.2.0
pillow==11.2.1
sqlparse==0.5.3
typing_extensions==4.13.2
